feat：mlapo; rope; prefetch; allgather; linear nz (#94)
* dsv3 support mlapo fusion operation.

* feat: add npu_interleave_rope for dsv3

* feat: add feature mla_ag_after_qlora

* feat: o proj,gmm,mlapo NZ;add p batch size; remove cast; o and mlapo prefetch;

+++ python/sglang/srt/layers/moe/ep_moe/layer.py
@@ -823,7 +823,7 @@ class DeepEPMoE(EPMoE):
             hidden_states = torch_npu.npu_grouped_matmul(
                 x=[hidden_states],
                 weight=[self.w2_weight],
-                scale=[self.w2_weight_scale.to(output_dtype)],
+                scale=[self.w2_weight_scale],
                 per_token_scale=[swiglu_out_scale],
                 split_item=2,
                 group_list_type=group_list_type,
